split the logic
