+++ vllm_rbln/v1/worker/optimum_model_runner.py
@@ -176,6 +176,8 @@ class RBLNOptimumModelRunner(LoRAModelRunnerMixin):
             # It is a temporary solution.
             if isinstance(self.model, RBLNOptimumDictTableMixin):
                 self.model.clear_dict_table()
+            if getattr(self.model, "attention_manager", None):
+                self.model.attention_manager.clear_dict_table()
             # Return empty ModelRunnerOutput if there's no work to do.
             return EMPTY_MODEL_RUNNER_OUTPUT
         # Prepare the decoder inputs.
